requirements.txt
# GitHub API
PyGithub==2.1.1

# MongoDB
pymongo==4.6.1
pymongoarrow==1.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Environment & Configuration
python-dotenv==1.0.0

# Scheduling
schedule==1.2.0

# HTTP Requests (for additional API calls)
requests==2.31.0

# Date/Time handling
python-dateutil==2.8.2

# Data Visualization (optional, for analysis)
matplotlib==3.8.2
plotly==5.18.0

# Logging
loguru==0.7.2

# Type hints
pydantic==2.5.3

# CLI and display
tabulate==0.9.0